# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace API",
        "description": "Marketplace API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sellers/get": {
            "get": {
                "tags": [
                    "Продавцы"
                ],
                "summary": "Получить список продавцов",
                "operationId": "7b268ca9fc30827540481bbf0c316d2d",
                "responses": {
                    "200": {
                        "description": "Список продавцов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Seller"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/sellers/create": {
            "post": {
                "tags": [
                    "Продавцы"
                ],
                "summary": "Добавить продавцов",
                "operationId": "50b72aa27952de3cb805fa192954560a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Имя продавца",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email продавца",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "Телефон продавца в формате 77001234567",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Пароль продавца",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно"
                    },
                    "422": {
                        "description": "Ошибка проверки"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/sellers/{id}/update": {
            "put": {
                "tags": [
                    "Продавцы"
                ],
                "summary": "Обновить данные продавца",
                "operationId": "b1bfc0e45226a12d00bc3d0166c9166c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продавца",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Имя продавца",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email продавца",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "Телефон продавца",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно обновлено"
                    },
                    "404": {
                        "description": "Продавец не найден"
                    },
                    "422": {
                        "description": "Ошибка проверки"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/sellers/{id}/delete": {
            "delete": {
                "tags": [
                    "Продавцы"
                ],
                "summary": "Удалить продавца",
                "operationId": "dc47ca709386e874889188a58d06f145",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продавца",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продавец успешно удален"
                    },
                    "404": {
                        "description": "Продавец не найден"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Seller": {
                "title": "Seller",
                "description": "Схема продавца",
                "properties": {
                    "id": {
                        "description": "ID продавца",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Имя продавца",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email продавца",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Телефон продавца в формате 77001234567",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}